Hack ctext to make highlighted text bold; syntax.tcl now ctext-aware
This hack to ctext is terrible, but moves a step toward the destination.
Currently, anything that gets highlighted gets forced to bold.  What we
-want- is to be able to specify stylings besides colors to ctext, so we
can pick bold or italic or maybe even a different font.

Syntax.tcl gets a complete gutting and gets re-implemented in terms
of ctext's highlighting, which is faster and better, modulo the bold
problem.

Also we've solarized the colors, which also too needs a better fix
than this, probably in colors.tcl.